use instancing draw calls
I was expecting a bigger performance increase than I got with that. The
project I'm using this library for still can't hit 60fps on Safari.

A potential alternative is to modify the sprite vertices with the
transformers in Javascript and use an indexed vertice array. This is the
more traditional way 2d sprite engines are done and may possibly have a
performance advantage over instancing.